# backend/src/main/java/com/spectemus/simul/backend/feedback/FeedbackReportExportResponse.java
package com.spectemus.simul.backend.feedback;

import java.time.Instant;
import java.util.List;

public record FeedbackReportExportResponse(
        Instant exportedAt,
        int count,
        List<FeedbackReport> reports) {}