# backend/src/main/java/com/spectemus/simul/backend/feedback/FeedbackReportListResponse.java
package com.spectemus.simul.backend.feedback;

import java.time.Instant;
import java.util.List;

public record FeedbackReportListResponse(
        Instant listedAt,
        int count,
        List<FeedbackReportSummary> reports) {}